Call `correct_float` in `simplify`
The strategy of only calling `correct_float` in lossy operations doesn't
quite work out, because the errors build up over time - just calling it
from (co)sine didn't fix the "sin(x)^2 + cos(x)^2" problem.

Instead, we'll just call it aggressively and hope the accuracy loss
isn't too much of an issue. The ideal (but high-effort solution) is to
tag `Decimal`s as "low precision" if an operation like sine was done on
them - maybe something for the future!